tls: Merge tls_xxx into pthread_keyxxx
it's always better to provide the standard api directly
