+++ libs/libc/pthread/pthread_setspecific.c
@@ -25,7 +25,9 @@
 #include <nuttx/config.h>
 
 #include <pthread.h>
+#include <assert.h>
 
+#include <nuttx/arch.h>
 #include <nuttx/tls.h>
 
 #if CONFIG_TLS_NELEM > 0
@@ -72,9 +74,23 @@
 
 int pthread_setspecific(pthread_key_t key, FAR const void *value)
 {
-  int ret = tls_set_value((int)key, (uintptr_t)value);
-  return ret < 0 ? -ret : 0;
+  FAR struct tls_info_s *info;
+
+  DEBUGASSERT(key >= 0 && key < CONFIG_TLS_NELEM);
+  if (key >= 0 && key < CONFIG_TLS_NELEM)
+    {
+      /* Get the TLS info structure from the current threads stack */
+
+      info = up_tls_info();
+      DEBUGASSERT(info != NULL);
+
+      /* Set the element value int the TLS info. */
+
+      info->tl_elem[key] = (uintptr_t)value;
+      return OK;
+    }
+
+  return EINVAL;
 }
 
 #endif /* CONFIG_TLS_NELEM */
-
