+++ libs/libc/tls/Make.defs
@@ -20,13 +20,15 @@
 
 CSRCS += task_getinfo.c tls_getinfo.c
 
+ifneq ($(CONFIG_TLS_TASK_NELEM),0)
+CSRCS += task_tls.c
+endif
+
 ifneq ($(CONFIG_TLS_NELEM),0)
-CSRCS += tls_alloc.c tls_free.c
-CSRCS += tls_setvalue.c tls_getvalue.c tls_destruct.c
+CSRCS += tls_destruct.c
 endif
 
 # Include tls build support
 
 DEPPATH += --dep-path tls
 VPATH += :tls
-
